Improve parsing of `${` template strings
This is incomplete, but still an improvement.

Related to https://github.com/Microsoft/tolerant-php-parser/issues/214

Note: The function never gets TokenKind::StringVarName for
OpenBraceDollarToken (`{$`), so it always uses parseExpression.
Also, TokenKind::Name is already used for parseExpression.

Add test cases for `"${var['stringIndex']}"`, as well of tests for
things tolerant-php-parser already supported due to parseExpression()

stringLiteral16.php was not used due to already being in skipped.json
(bug for making skipped.json specific was filed)

For issues #219 and #214