Workaround for #185 when running tests locally
This is annoying for developers running unit tests
with short_open_tag enabled
(e.g. if you build php from source with support for that)

- By making this change, this PR ensures that those developers
  don't accidentally commit modified versions of
  programStructure13.php.tree, .diag, etc.

The short_open_tag option cannot be changed at runtime.

An alternative could be to add short_open_tag=0 to phpunit.xml
(But in the future, someone might want to explicitly test the
behavior with short open tags?)