Remove a TODO that no longer applies in php 8.0
Any expression can be used as of 8.0 without an unexpected syntax error.

In encapsulated strings,
`"${a->prop}"` is equivalent to `"{${(a->prop)}}"` in php 8.0 and a
syntax error in previous php versions.

(Get the property `prop` of the global constant `a` (constants can be
objects in php 8.1), then get the variable with the name equal to that
string)